backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
beautifulsoup4
firebase-admin
google-generativeai
python-dotenv
numpy
psycopg2-binary
feedparser
pydantic
